apivenv/drfdjango/requirements.txt
anyio==3.6.2
asgiref==3.6.0
click==8.1.3
colorama==0.4.6
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
fastapi==0.95.1
h11==0.14.0
idna==3.4
Markdown==3.4.3
pydantic==1.10.7
pytz==2023.3
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
uvicorn==0.22.0
